+++ spec/rubocop/cop/rails/skips_model_validations_spec.rb
@@ -21,8 +21,30 @@ describe RuboCop::Cop::Rails::SkipsModelValidations, :config do
   let(:msg) { 'Avoid using `%s` because it skips validations.' }
   let(:cop_config) { cop_config }
 
+  methods_with_arguments = described_class::METHODS_WITH_ARGUMENTS
+
   context 'with default blacklist' do
     cop_config['Blacklist'].each do |method_name|
+      it "registers an offense for `#{method_name}`" do
+        inspect_source(cop, "User.#{method_name}(:attr)")
+        expect(cop.offenses.size).to eq(1)
+        expect(cop.messages)
+          .to eq([format(msg, method_name)])
+      end
+    end
+  end
+
+  context 'with methods that require at least an argument' do
+    methods_with_arguments.each do |method_name|
+      it "doesn't register an offense for `#{method_name}`" do
+        inspect_source(cop, "User.#{method_name}")
+        expect(cop.offenses).to be_empty
+      end
+    end
+  end
+
+  context "with methods that don't require an argument" do
+    (cop_config['Blacklist'] - methods_with_arguments).each do |method_name|
       it "registers an offense for `#{method_name}`" do
         inspect_source(cop, "User.#{method_name}")
         expect(cop.offenses.size).to eq(1)